added tracking funtionality for custom control
